==> preparacion-datos-scripts/11_BrainLat_ImputacionDatos_soloIFSyAges.Rmd <==
---
title: 'BrainLat: Imputando Datos (IFS y Age igual a 0)'
author: "Yassmina"
date: "2025-05-21"
output: html_notebook
---

# BRAINLAT: IMPUTANDO DATOS (IFS Y AGE IGUAL A 0).

```{r setup, include=FALSE}
library(svglite)
knitr::opts_knit$set(root.dir = "directorio_censurado")
```

## 1. Lectura y preparación datos:
```{r}
library(dplyr)

data = read.csv("BrainLat_Filtrado2.csv", sep = ";")
data$sex <- as.factor(data$sex)
data$laterality <- as.factor(data$laterality)
data$Age <- as.integer(data$Age)

data$T1 <- as.factor(data$T1)
data$Rest <- as.factor(data$Rest)
data$DWI <- as.factor(data$DWI)
data$MF <- as.factor(data$MF)
data$eeg <- as.factor(data$eeg)

summary(data)
```
### 1.1. Igualando a "NA" los Age igual a 0: 
```{r}
data$Age[which(data$Age == 0)] <- NA
```

## 2. Imputación datos:
```{r}
library(mice)

ifs_cols <- setdiff(grep("^ifs_", names(data), value = TRUE), "ifs_total_score")

# Función para verificar y corregir los totales
corregir_totales <- function(data, sub_cols, total_col) {
  data$suma_correcta <- rowSums(data[, sub_cols], na.rm = TRUE)     # Sumar valores presentes
  data[, total_col][data[, total_col] != data$suma_correcta] <- NA  # Borrar total incorrecto
  data$suma_correcta <- NULL                                        # Eliminar columna auxiliar
  return(data)
}

data <- corregir_totales(data, ifs_cols, "ifs_total_score")

variables_imputar <- c("Age", "sex", "years_education", "laterality", "country", ifs_cols, "ifs_total_score", "mini_sea_fer", "mini_sea_tom", "MF")

# Configuramos relaciones en mice
formulas <- make.formulas(data[, variables_imputar])
formulas$ifs_total_score <- as.formula(paste("ifs_total_score ~", paste(ifs_cols, collapse = " + ")))
post <- mice::make.post(data[, variables_imputar])
post["Age"] <- "imp[[j]][, i] <- ifelse(imp[[j]][, i] < 5, NA, imp[[j]][, i])"

mice_imputado <- mice(data[, variables_imputar], method = "pmm", formulas = formulas, post = post, m = 5, maxit = 10, seed = 123)
```
```{r}
library(ggplot2)
library(gridExtra)
library(mice)

imputed_data <- complete(mice_imputado, action = "long")

plots <- list()

for (m in 1:5) {
  imputacion_m <- subset(imputed_data, .imp == m)
  
  for (var in variables_imputar) {
    if (is.numeric(data[[var]])) {
      plot_data <- data.frame(
        value = c(data[[var]], imputacion_m[[var]]),
        source = c(
          rep("Original", nrow(data)),
          rep(paste("Imputación", m), nrow(imputacion_m))
        )
      )
      
      p <- ggplot(plot_data, aes(x = source, y = value, fill = source)) +
        geom_boxplot() +
        labs(title = paste(var, "- Imputación", m), x = "", y = "Value") +
        theme_minimal() +
        theme(axis.text.x = element_text(angle = 45, hjust = 1),
              legend.position = "none")
      
      plots[[paste(var, "imp", m, sep = "_")]] <- p
    }
  }
}


if (length(plots) >= 1) {
  grid.arrange(grobs = plots[1:min(6, length(plots))], ncol = 2)
}

pdf("comparacion_imputaciones3.pdf", width = 11, height = 8)
for (i in seq(1, length(plots), by = 6)) {
  grid.arrange(grobs = plots[i:min(i + 5, length(plots))], ncol = 2)
}
dev.off()
```

```{r}
data_imputado <- complete(mice_imputado, m = 2)
data_imputado <- data_imputado[["1"]]
```


## 3. Corrección errores restantes:
```{r}
for (i in 1:nrow(data)) {
  data_imputado$ifs_total_score[i] <- sum(data_imputado[i, ifs_cols], na.rm = TRUE)
}
```

## 4. Juntado columnas no imputadas:
```{r}
data_imputado <- cbind(data[, !colnames(data) %in% colnames(data_imputado)], data_imputado)
data_imputado <- data_imputado[, colnames(data)]
summary(data_imputado)
```

## 4. Guardando datos:
```{r}
# creación de center
data_imputado$center <- gsub("^sub-([A-Za-z]{2,3}).*", "\\1", data_imputado$MRI_ID)
data_imputado <- data_imputado %>% relocate(center, .after = 8)  # La colocamos en la posición 9
str(data_imputado[, 1:10])
write.csv(data_imputado, "BrainLat_Imputado3.csv", row.names = FALSE)
```
